Python/requirements.txt
easydict
fire
numpy
onnxruntime
opencv-python
PyYAML